requirements.txt
gradio>=4.19.2
openai>=1.13.3
google-auth-oauthlib>=1.2.0
google-auth>=2.28.1
google-api-python-client>=2.120.0
python-dotenv>=1.0.1
pandas>=2.2.1
numpy>=1.26.4
pytest>=8.0.2
pytest-asyncio>=0.23.5
black>=24.2.0
ruff>=0.3.0 